Implement authentication middleware, and add user role access control

+++ controllers/user/deleteUser.js
@@ -1,8 +1,13 @@
-const User = require("../../models/user");
+const User = require("../../models/user.model");
 
 const deleteUser = async (req, res)=>{
     try{
         const {id} = req.params;
+
+        //safety check - admin can not delete itself
+        if (req.user.id === id) {
+            return res.status(400).json({ message: 'Cannot delete yourself' })
+        }
         
         const user = await User.findByIdAndDelete(id);
         if(!user) {
